chore: rename to asciicast.js
