+++ web_src/js/markup/content.js
@@ -1,7 +1,7 @@
 import {renderMermaid} from './mermaid.js';
 import {renderMath} from './math.js';
 import {renderCodeCopy} from './codecopy.js';
-import {renderAsciinemaPlayer} from './asciinema.js';
+import {renderAsciinemaPlayer} from './asciicast.js';
 import {initMarkupTasklist} from './tasklist.js';
 
 // code that runs for all markup content
